requirements.txt
langchain
langgraph
google-generativeai
requests
beautifulsoup4
fpdf
python-dotenv
streamlit
